requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-cors-headers==3.10.0
django-filter==21.1
django-rest-auth
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
docx==0.2.4
docxcompose==1.3.3
docxtpl==0.14.2
idna==3.3
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.0.3
lxml==4.6.4
MarkupSafe==2.0.1
oauthlib==3.1.1
openapi-codec==1.3.2
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.20
python-docx==0.8.11
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
